Adjust gain header if header is complete, even if entire frame isn't This avoids situations where adding a tag fools the gain adjustment function into thinking the final frame is complete, when it wasn't earlier.
Also, little tidbits to stop Visual Studio 2005 from complaining

+++ mp3gain/gain_analysis.c
@@ -302,7 +302,7 @@ AnalyzeSamples ( const Float_t* left_samples, const Float_t* right_samples, size
         return GAIN_ANALYSIS_OK;
 
     cursamplepos = 0;
-    batchsamples = num_samples;
+    batchsamples = (long)num_samples;
 
     switch ( num_channels) {
     case  1: right_samples = left_samples;
